validate venture and timeline names for duplicates
