+++ src/component/app/RegisteredUserRoute.tsx
@@ -5,20 +5,18 @@ import { Begin } from "component/begin";
 import { Home } from "component/page/home";
 import JoinVenture from "component/page/joinventure";
 import { UserContext } from "context/UserContext";
+import { VentureContext } from "context/VentureContext";
+import { useVentures } from "module/hook/ui/useVentures";
 
 import { VentureRoutes } from "./VentureRoutes";
 
 export function RegisteredUserRoute() {
   const { user, status } = useContext(UserContext);
 
-  if (status === "loading") {
-    return <span>Loading user...</span>;
-  }
-
   const returnTo = window.location.pathname + window.location.search;
   if (status === "error") {
     return <Navigate to="/signin" state={{ returnTo }} />;
-  } else if (!user) {
+  } else if (status === "success" && !user) {
     return <Navigate to="/profile" state={{ returnTo }} />;
   }
 
@@ -28,9 +26,22 @@ export function RegisteredUserRoute() {
       <Route path="begin" element={<Begin />} />
       <Route path="editprofile" element={<EditProfile />} />
       <Route path="newventure" element={<NewVenture />} />
-      <Route path="/*" element={<VentureRoutes />} />
       <Route path=":ventureSlug/*" element={<VentureRoutes />} />
+      <Route path="*" element={<VentureRoutes />} />
     </Routes>
+  );
+}
+
+function NewVenture() {
+  // We provide this in VentureRoutes, but /newventure is rendered outside of this context
+  // so we have to provide it separately here.
+  const ventureContext = useVentures("");
+  const variantType = "isVenture";
+  const isActive = "settings";
+  return (
+    <VentureContext.Provider value={ventureContext}>
+      <Home variantType={variantType} isActive={isActive} />
+    </VentureContext.Provider>
   );
 }
 
@@ -38,10 +49,4 @@ function EditProfile() {
   const modalType = "editProfile";
   const isVisible = "showModal";
   return <Home modalType={modalType} isVisible={isVisible} />;
-}
-
-export function NewVenture() {
-  const variantType = "isVenture";
-  const isActive = "settings";
-  return <Home variantType={variantType} isActive={isActive} />;
 }
