+++ src/component/app/VentureRoutes.tsx
@@ -1,73 +1,21 @@
 import { useContext, useEffect } from "react";
 import TagManager from "react-gtm-module";
-import { Navigate, Route, Routes, useParams } from "react-router";
+import { Navigate, Route, Routes } from "react-router";
+import { useParams } from "react-router-dom";
 
 import { Home } from "component/page/home";
-import { AuthContext } from "context/AuthContext";
 import { UserContext } from "context/UserContext";
-import { IVentureContext, VentureContext } from "context/VentureContext";
-import { calculateNamedSlug, getUniqueListBy } from "module/helpers";
-import { useRoleByTimelines, useRoleByVentures } from "module/hook/role";
-import { useTimelinesByUserId } from "module/hook/timeline";
-import { useVentureMembers } from "module/hook/user";
-import { useVentureByTimeline, useVenturesByUser } from "module/hook/venture";
-import { IRole } from "module/interface/role";
-import { ITimeline } from "module/interface/timeline";
-import { IUser, UserRole } from "module/interface/user";
-import { IVenture } from "module/interface/venture";
+import { VentureContext } from "context/VentureContext";
+import { calculateNamedSlug } from "module/helpers";
+import { useVentures } from "module/hook/ui/useVentures";
 
 import { TimelineRoutes } from "./TimelineRoutes";
 
-function injectVentureRoles(
-  ventures: IVenture[],
-  userId: string | undefined,
-  ventureRoles: IRole[]
-) {
-  return (
-    ventures.map((venture) => {
-      const ventureRole = ventureRoles.find(
-        (role: IRole) =>
-          role.subjectId === userId && role.ventureId === venture.id
-      );
-      if (!ventureRole) return venture;
-      return {
-        ...venture,
-        userRole: ventureRole.role as UserRole,
-      };
-    }) || []
-  );
-}
-
-function injectTimelineRoles(
-  timelines: ITimeline[],
-  userId: string | undefined,
-  ventureRoles: IRole[],
-  timelineRoles: IRole[]
-) {
-  return (
-    timelines.map((timeline) => {
-      const ventureRole = ventureRoles.find(
-        (role) =>
-          role.subjectId === userId && role.ventureId === timeline.ventureId
-      );
-      const timelineRole = timelineRoles.find(
-        (role) => role.subjectId === userId && role.timelineId === timeline.id
-      );
-      const userRole = timelineRole?.role || ventureRole?.role;
-      if (!userRole) return timeline;
-      return {
-        ...timeline,
-        userRole: userRole as UserRole,
-      };
-    }) || []
-  );
-}
-
 export function VentureRoutes() {
-  const { user } = useContext(UserContext);
-  const { token } = useContext(AuthContext);
+  const { user: { id: userId } = { id: undefined } } = useContext(UserContext);
   const { ventureSlug } = useParams();
-  const userId = user?.id;
+  const ventureContext = useVentures(ventureSlug);
+  const { loading, ventures, currentVenture, timelines } = ventureContext;
 
   useEffect(() => {
     if (userId) {
@@ -81,113 +29,13 @@ export function VentureRoutes() {
     }
   }, [userId]);
 
-  const { data: venturesByUserData = [], status: venturesByUserStatus } =
-    useVenturesByUser({
-      userId,
-      token,
-    });
-
-  const { data: timelinesByUserData = [], status: timelinesByUserStatus } =
-    useTimelinesByUserId({
-      userId,
-      token,
-    });
-
-  const {
-    data: venturesByTimelineData = [],
-    status: venturesByTimelineStatus,
-  } = useVentureByTimeline({
-    ventureIds: [...new Set(timelinesByUserData?.map((t) => t.ventureId))],
-    token,
-  });
-
-  const combinedVentures = getUniqueListBy(
-    [...venturesByTimelineData, ...venturesByUserData],
-    "id"
-  );
-
-  const { data: ventureRolesData = [], status: ventureRolesStatus } =
-    useRoleByVentures({
-      ventures: combinedVentures,
-      token,
-    });
-
-  const ventures = injectVentureRoles(
-    combinedVentures,
-    userId,
-    ventureRolesData
-  );
-  ventures.sort((a, b) =>
-    calculateNamedSlug(a).localeCompare(calculateNamedSlug(b))
-  );
-
-  const { data: timelineRolesData = [], status: timelineRolesStatus } =
-    useRoleByTimelines({
-      timelines: timelinesByUserData,
-      token,
-    });
-
-  const currentVenture = ventures.find(
-    (v) => calculateNamedSlug(v) === ventureSlug
-  );
-
-  const {
-    data: currentVentureUsers = [],
-    status: currentVentureMembersStatus,
-  } = useVentureMembers({
-    ventureId: currentVenture?.id,
-    token,
-  });
-
-  const loading =
-    venturesByUserStatus === "loading" ||
-    venturesByTimelineStatus === "loading" ||
-    ventureRolesStatus === "loading" ||
-    timelinesByUserStatus === "loading" ||
-    timelineRolesStatus === "loading" ||
-    currentVentureMembersStatus === "loading";
-
-  const timelines = injectTimelineRoles(
-    timelinesByUserData,
-    userId,
-    ventureRolesData,
-    timelineRolesData
-  );
-
-  const venturesLoaded =
-    venturesByUserStatus === "success" &&
-    venturesByTimelineStatus === "success";
-  if (venturesLoaded && (!ventureSlug || !currentVenture)) {
+  if (!loading && (!ventureSlug || !currentVenture)) {
     if (ventures.length > 0) {
       return <Navigate replace to={`/${calculateNamedSlug(ventures[0])}`} />;
     } else {
       return <Navigate replace to="/begin" />;
     }
   }
-
-  const currentVentureTimelines = timelines.filter(
-    (t) => t.ventureId === currentVenture?.id
-  );
-
-  const currentVentureMembers = currentVentureUsers
-    .map((user) => ({
-      user,
-      role: ventureRolesData.find(
-        (r) => r.ventureId === currentVenture?.id && r.subjectId === user.id
-      ),
-    }))
-    .filter((member): member is { user: IUser; role: IRole } =>
-      Boolean(member.role)
-    );
-
-  const ventureContext: IVentureContext = {
-    ventures,
-    timelines,
-
-    currentVenture,
-    currentVentureTimelines,
-    currentVentureMembers,
-  };
 
   const hasTimelines = timelines.some(
     (t) => t.ventureId === currentVenture?.id
