+++ src/module/hook/ui/useVentures.ts
@@ -1,0 +1,165 @@
+import { useContext } from "react";
+
+import { AuthContext } from "../../../context/AuthContext";
+import { UserContext } from "../../../context/UserContext";
+import { IVentureContext } from "../../../context/VentureContext";
+import { calculateNamedSlug, getUniqueListBy } from "../../helpers";
+import { useRoleByTimelines, useRoleByVentures } from "../../hook/role";
+import { useTimelinesByUserId } from "../../hook/timeline";
+import { useVentureMembers } from "../../hook/user";
+import { useVentureByTimeline, useVenturesByUser } from "../../hook/venture";
+import { IRole } from "../../interface/role";
+import { ITimeline } from "../../interface/timeline";
+import { IUser, UserRole } from "../../interface/user";
+import { IVenture } from "../../interface/venture";
+
+function injectVentureRoles(
+  ventures: IVenture[],
+  userId: string | undefined,
+  ventureRoles: IRole[]
+) {
+  return (
+    ventures.map((venture) => {
+      const ventureRole = ventureRoles.find(
+        (role: IRole) =>
+          role.subjectId === userId && role.ventureId === venture.id
+      );
+      if (!ventureRole) return venture;
+      return {
+        ...venture,
+        userRole: ventureRole.role as UserRole,
+      };
+    }) || []
+  );
+}
+
+function injectTimelineRoles(
+  timelines: ITimeline[],
+  userId: string | undefined,
+  ventureRoles: IRole[],
+  timelineRoles: IRole[]
+) {
+  return (
+    timelines.map((timeline) => {
+      const ventureRole = ventureRoles.find(
+        (role) =>
+          role.subjectId === userId && role.ventureId === timeline.ventureId
+      );
+      const timelineRole = timelineRoles.find(
+        (role) => role.subjectId === userId && role.timelineId === timeline.id
+      );
+      const userRole = timelineRole?.role || ventureRole?.role;
+      if (!userRole) return timeline;
+      return {
+        ...timeline,
+        userRole: userRole as UserRole,
+      };
+    }) || []
+  );
+}
+
+export function useVentures(ventureSlug: string): IVentureContext {
+  const { user } = useContext(UserContext);
+  const { token } = useContext(AuthContext);
+  const userId = user?.id;
+
+  const { data: venturesByUserData = [], status: venturesByUserStatus } =
+    useVenturesByUser({
+      userId,
+      token,
+    });
+
+  const { data: timelinesByUserData = [], status: timelinesByUserStatus } =
+    useTimelinesByUserId({
+      userId,
+      token,
+    });
+
+  const {
+    data: venturesByTimelineData = [],
+    status: venturesByTimelineStatus,
+  } = useVentureByTimeline({
+    ventureIds: [...new Set(timelinesByUserData?.map((t) => t.ventureId))],
+    token,
+  });
+
+  const combinedVentures = getUniqueListBy(
+    [...venturesByTimelineData, ...venturesByUserData],
+    "id"
+  );
+
+  const { data: ventureRolesData = [], status: ventureRolesStatus } =
+    useRoleByVentures({
+      ventures: combinedVentures,
+      token,
+    });
+
+  const ventures = injectVentureRoles(
+    combinedVentures,
+    userId,
+    ventureRolesData
+  );
+  ventures.sort((a, b) =>
+    calculateNamedSlug(a).localeCompare(calculateNamedSlug(b))
+  );
+
+  const { data: timelineRolesData = [], status: timelineRolesStatus } =
+    useRoleByTimelines({
+      timelines: timelinesByUserData,
+      token,
+    });
+
+  const currentVenture = ventures.find(
+    (v) => calculateNamedSlug(v) === ventureSlug
+  );
+
+  const {
+    data: currentVentureUsers = [],
+    status: currentVentureMembersStatus,
+  } = useVentureMembers({
+    ventureId: currentVenture?.id,
+    token,
+  });
+
+  const loading =
+    venturesByUserStatus !== "success" ||
+    venturesByTimelineStatus !== "success" ||
+    ventureRolesStatus !== "success" ||
+    timelinesByUserStatus !== "success" ||
+    timelineRolesStatus !== "success" ||
+    (currentVentureMembersStatus !== "success" && Boolean(ventureSlug));
+
+  const timelines = injectTimelineRoles(
+    timelinesByUserData,
+    userId,
+    ventureRolesData,
+    timelineRolesData
+  );
+
+  const currentVentureTimelines = timelines.filter(
+    (t) => t.ventureId === currentVenture?.id
+  );
+
+  const currentVentureMembers = currentVentureUsers
+    .map((user) => ({
+      user,
+      role: ventureRolesData.find(
+        (r) => r.ventureId === currentVenture?.id && r.subjectId === user.id
+      ),
+    }))
+    .filter((member): member is { user: IUser; role: IRole } =>
+      Boolean(member.role)
+    );
+
+  const ventureContext: IVentureContext = {
+    loading,
+    ventures,
+    timelines,
+
+    currentVenture,
+    currentVentureTimelines,
+    currentVentureMembers,
+  };
+
+  return ventureContext;
+}
